+++ src/main/java/com/sideProject/PlanIT/domain/post/entity/ENUM/PostStatus.java
@@ -1,0 +1,7 @@
+package com.sideProject.PlanIT.domain.post.entity.ENUM;
+
+public enum PostStatus {
+    NOT_STARTED,
+    IN_PROGRESS,
+    EXPIRED
+}
