+++ src/main/java/com/sideProject/PlanIT/domain/schedule/entity/ENUM/ScheduleAttendance.java
@@ -1,4 +1,4 @@
-package com.sideProject.PlanIT.domain.schedule.ENUM;
+package com.sideProject.PlanIT.domain.schedule.entity.ENUM;
 
 public enum ScheduleAttendance {
     ATTNENDED,
