+++ src/main/java/com/sideProject/PlanIT/domain/user/entity/ENUM/MemberRole.java
@@ -2,5 +2,6 @@ package com.sideProject.PlanIT.domain.user.entity.ENUM;
 
 public enum MemberRole {
     MEMBER,
-    TRAINER
+    TRAINER,
+    ADMIN
 }
